src/requirements.txt
langchain 
langchain-community 
langchain-google-genai
langchain_openai
langgraph
python-dotenv
pydantic
duckduckgo-search
beautifulsoup4
scikit-learn